requirements.txt
absl-py==1.0.0
anyio==3.3.2
appnope==0.1.2
argon2-cffi==21.1.0
astunparse==1.6.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
blis==0.7.4
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
click==7.1.2
cloudpickle==2.0.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.24
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
flatbuffers==2.0
gast==0.4.0
glfw==2.3.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
# Editable install with no version control (gym==0.10.5)
-e /Users/rugvedmavidipalli/Desktop/classes/Artificial Inteligence/venv/lib/python3.9/site-packages
h5py==3.6.0
idna==3.2
imageio==2.10.1
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.0.1
jupyter-client==7.0.5
jupyter-core==4.8.1
jupyter-server==1.11.0
jupyterlab==3.1.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lockfile==0.12.2
lxml==4.6.4
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.5
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pluggy==1.0.0
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.19.1
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pygame==2.0.1
pyglet==1.5.21
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.8
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
spacy==3.1.2
spacy-legacy==3.0.8
srsly==2.4.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
Theano==1.0.5
thinc==8.0.10
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
typer==0.3.2
typing-extensions==3.10.0.2
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
